refactor: Extract http_service::post()
