+++ src/main.rs
@@ -2,6 +2,7 @@ extern crate core;
 
 mod commands;
 mod git_service;
+mod http_service;
 mod pipelines;
 mod release;
 mod semantic_version;
