+++ src/pipelines/gitlab_ci.rs
@@ -1,6 +1,7 @@
-use crate::git_service;
 use crate::pipelines::Pipeline;
 use crate::release::Release;
+use crate::{git_service, http_service};
+use reqwest::header::HeaderMap;
 use serde_json::Value;
 use std::collections::HashMap;
 
@@ -38,7 +39,15 @@ impl Pipeline for GitlabCI {
     }
 
     fn create_release(&self, release: &Release) -> HashMap<String, Value> {
-        let path = format!("projects/{}/releases", self.env_var("CI_PROJECT_ID"));
+        let url = format!(
+            "{}/projects/{}/releases",
+            self.env_var("CI_API_V4_URL"),
+            self.env_var("CI_PROJECT_ID")
+        );
+
+        let mut headers = HeaderMap::new();
+        headers.insert("JOB-TOKEN", self.env_var("CI_JOB_TOKEN").parse().unwrap());
+
         let mut body = HashMap::new();
         body.insert("name", release.name.clone());
         body.insert("description", release.description.clone());
@@ -46,7 +55,7 @@ impl Pipeline for GitlabCI {
         body.insert("tag_message", release.tag_message.clone());
         body.insert("ref", self.env_var("CI_COMMIT_SHA"));
 
-        self.api_v4_post(path, body)
+        http_service::post(url, Some(headers), Some(body))
     }
 }
 
@@ -54,31 +63,5 @@ impl GitlabCI {
     fn git_origin_pushurl(&self, url: String) {
         git_service::set_config_value("remote.origin.pushurl", &format!("{}.git", url))
             .unwrap_or_else(|e| panic!("{}", e));
-    }
-
-    fn api_v4_post(&self, path: String, body: HashMap<&str, String>) -> HashMap<String, Value> {
-        let url = format!("{}/{}", self.env_var("CI_API_V4_URL"), path);
-        let client = reqwest::blocking::Client::new();
-
-        let response = client
-            .post(url)
-            .header("PRIVATE-TOKEN", self.env_var("DEV_TOKEN"))
-            .json(&body)
-            .send()
-            .unwrap_or_else(|e| panic!("{}", e));
-
-        let status = response.status();
-        if status.is_success() {
-            response
-                .json::<HashMap<String, Value>>()
-                .unwrap_or_else(|e| panic!("{}", e))
-        } else {
-            let headers = response.headers().clone();
-            let body = response.text().unwrap();
-            panic!(
-                "Status: {}\nHeaders:\n{:#?}\nBody:\n{}",
-                status, headers, body
-            )
-        }
     }
 }
